train/requirements-train.txt
# Python 3.10+ recommended. Install in a venv (not the minimal API venv).
torch>=2.1.0
torchvision>=0.16.0
onnx>=1.15.0
Pillow>=10.0.0
